Move sso auth logic to auth folder
