+++ assets/js/lib/auth/config.test.js
@@ -1,0 +1,7 @@
+import { isSingleSignOnEnabled } from './config';
+
+describe('auth config', () => {
+  it('should check if single sign on is enabled', () => {
+    expect(isSingleSignOnEnabled()).toBe(false);
+  });
+});
