+++ assets/js/pages/Users/CreateUserPage.test.jsx
@@ -13,7 +13,7 @@ import { faker } from '@faker-js/faker';
 import * as router from 'react-router';
 
 import { networkClient } from '@lib/network';
-import * as usersModel from '@lib/model/users';
+import * as authConfig from '@lib/auth/config';
 import { abilityFactory, userFactory } from '@lib/test-utils/factories/users';
 
 import CreateUserPage from './CreateUserPage';
@@ -147,7 +147,7 @@ describe('CreateUserPage', () => {
 
   describe('Single sign on', () => {
     it('should redirect to not found page', async () => {
-      jest.spyOn(usersModel, 'isSingleSignOnEnabled').mockReturnValue(true);
+      jest.spyOn(authConfig, 'isSingleSignOnEnabled').mockReturnValue(true);
 
       render(<CreateUserPage />);
